pyproject.toml
[tool.poetry]
name = "Ateneus Cooperatius"
version = "0.1.0"
description = ""
authors = ["Codi Cooperatiu <hola@codi.coop>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2"
psycopg2-binary = "2.7.5"
weasyprint = "^52.5"
django-summernote = "^0.8.20"
django-constance = {extras = ["database"], version = "^2.7.0"}
boto3 = "^1.19.9"
django-storages = "^1.12.3"
django-grappelli = "^2.15.3"
gunicorn = "^20.1.0"
easy-thumbnails = "^2.7.2"
django-modelclone = "^0.7.1"
openpyxl = "^3.0.9"
django-object-actions = "^3.0.2"
django-maintenance-mode = "^0.16.1"
django-mail-queue = "^3.2.4"
django-logentry-admin = "^1.1.0"
django-tagulous = "^1.3.0"
django-extensions = "^3.1.3"
django-q = "^1.3.9"
mailing-manager = "^0.1.0"
django-environ = "^0.8.1"
six = "^1.16.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
